requirements.txt
openai==0.27.4
pinecone-client==2.2.1
llama-index==0.5.10
mygene==3.2.2
biopython==1.81
colorama==0.4.6
tiktoken==0.3.0
backoff==2.2.1